+++ src/main/java/com/dhbackend/odontologia_integradorfinal/service/OdontologoService.java
@@ -1,5 +1,6 @@
 package com.dhbackend.odontologia_integradorfinal.service;
 
+import com.dhbackend.odontologia_integradorfinal.controller.exception.ResourceNotFoundException;
 import com.dhbackend.odontologia_integradorfinal.mapper.OdontologoMapper;
 import com.dhbackend.odontologia_integradorfinal.model.OdontologoDto;
 import com.dhbackend.odontologia_integradorfinal.persistence.entities.Odontologo;
@@ -78,8 +79,12 @@ public class OdontologoService {
     @Transactional
     public void deleteOdontologo(int id) {
         try {
+            if (!repository.existsById(id)) {
+                logger.error("No se encontro el odontologo con id: " + id);
+                throw new ResourceNotFoundException("No se encontro el odontologo con id: " + id);
+            }
             repository.deleteById(id);
-            logger.info("Se elimino el odontologo con id: " + id);
+            logger.info("Se eliminó el odontologo con id: " + id);
         } catch (Exception e) {
             logger.error("No se pudo eliminar el odontologo por: " + e.getMessage());
         }
